updated HomeController and LibraryService

+++ homework_3_new/librawherey1/src/main/java/com/example/librawherey1/model/Libraries.java
@@ -9,19 +9,19 @@ import javax.persistence.*;
 @Table(name = "libraries")
 public class Libraries {
     @Id
-    private Long id;
+    private Integer id;
     String name;
     String city;
     String street;
     String opening_hours;
     String phone;
-    Integer lat;
-    Integer lon;
+    Double lat;
+    Double lon;
 
     public Libraries() {
     }
 
-    public Libraries(Long id, String name, String city, String street, String opening_hours, String phone, Integer lat, Integer lon) {
+    public Libraries(Integer id, String name, String city, String street, String opening_hours, String phone, Double lat, Double lon) {
         this.id = id;
         this.name = name;
         this.city = city;
