+++ homework_3_new/librawherey1/src/main/java/com/example/librawherey1/model/Schools.java
@@ -10,19 +10,19 @@ import javax.persistence.*;
 
 public class Schools {
     @Id
-    private Long id;
+    private Integer id;
     String name;
     String city;
     String street;
     String opening_hours;
     String phone;
-    Integer lat;
-    Integer lon;
+    Double lat;
+    Double lon;
 
     public Schools() {
     }
 
-    public Schools(Long id, String name, String city, String street, String opening_hours, String phone, Integer lat, Integer lon) {
+    public Schools(Integer id, String name, String city, String street, String opening_hours, String phone, Double lat, Double lon) {
         this.id = id;
         this.name = name;
         this.city = city;
